refactor: global state order/cart/product deletion
